requirement.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
docarray==0.40.0
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.27
langchain-openai==0.2.13
langchain-text-splitters==0.3.4
langsmith==0.2.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai==1.58.1
openai-whisper==20240930
orjson==3.10.12
packaging==24.2
propcache==0.2.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.18.0
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
yt-dlp==2024.12.13
